+++ blocklets/ai-studio/src/components/file-editor/output/AddOutputVariableButton.tsx
@@ -8,28 +8,21 @@ import PlusIcon from '@iconify-icons/tabler/plus';
 import { Box, Divider, ListItemIcon, MenuItem } from '@mui/material';
 import { sortBy } from 'lodash';
 
-import useCallAgentOutput from '../use-call-agent-output';
+import useCallAgentOutput, { useCallAgentCustomOutputDialogState } from '../use-call-agent-output';
 import { runtimeOutputVariables } from './type';
 
 const ActionRefOutput = ({
-  onSelect,
-  onDeleteSelect,
   assistant,
   projectId,
   gitRef,
 }: {
-  onSelect?: (value: Partial<Omit<OutputVariableYjs, 'id'>>) => void;
-  onDeleteSelect?: (value: { id: string }) => void;
   assistant: AssistantYjs;
   projectId: string;
   gitRef: string;
 }) => {
   const { t } = useLocaleContext();
-  const { outputs, outputsExist, getRefOutputId, getOutputI18nName } = useCallAgentOutput({
-    projectId,
-    gitRef,
-    assistant,
-  });
+  const { outputs } = useCallAgentOutput({ projectId, gitRef, assistant });
+  const { onOpen } = useCallAgentCustomOutputDialogState(projectId, gitRef, assistant.id);
 
   if (!outputs?.length) return null;
 
@@ -39,28 +32,12 @@ const ActionRefOutput = ({
         {`${t('ref')}${t('outputs')}`}
       </Divider>
 
-      {outputs.map((output) => (
-        <MenuItem
-          key={output.id}
-          selected={outputsExist.has(output.id)}
-          onClick={(e) => {
-            e.stopPropagation();
-
-            if (outputsExist.has(output.id)) {
-              onDeleteSelect?.({ id: getRefOutputId(output.id) || '' });
-            } else {
-              onSelect?.({ name: output.name, from: { type: 'output', id: output.id } });
-            }
-          }}>
-          <ListItemIcon>
-            <Icon icon={BranchIcon} />
-          </ListItemIcon>
-          <Box flex={1}>{getOutputI18nName(output.name || '')}</Box>
-          <Box sx={{ width: 40, textAlign: 'right' }}>
-            {outputsExist.has(output.id) && <Box component={Icon} icon={CheckIcon} />}
-          </Box>
-        </MenuItem>
-      ))}
+      <MenuItem onClick={() => onOpen(true)}>
+        <ListItemIcon>
+          <Icon icon={PlusIcon} />
+        </ListItemIcon>
+        <Box flex={1}>{t('selectCustomOutput')}</Box>
+      </MenuItem>
     </>
   );
 };
@@ -69,7 +46,6 @@ export default function AddOutputVariableButton({
   allSelectAgentOutputs,
   assistant,
   onSelect,
-  onDeleteSelect,
   onSelectAll,
   projectId,
   gitRef,
@@ -77,7 +53,6 @@ export default function AddOutputVariableButton({
   allSelectAgentOutputs?: NonNullable<AssistantYjs['outputVariables']>[string]['data'][];
   assistant: AssistantYjs;
   onSelect?: (value: Partial<Omit<OutputVariableYjs, 'id'>>) => void;
-  onDeleteSelect?: (value: { id: string }) => void;
   onSelectAll?: (list: NonNullable<AssistantYjs['outputVariables']>[string]['data'][]) => void;
   projectId: string;
   gitRef: string;
@@ -169,13 +144,7 @@ export default function AddOutputVariableButton({
         </>
       )}
 
-      <ActionRefOutput
-        projectId={projectId}
-        gitRef={gitRef}
-        assistant={assistant}
-        onDeleteSelect={onDeleteSelect}
-        onSelect={onSelect}
-      />
+      <ActionRefOutput projectId={projectId} gitRef={gitRef} assistant={assistant} />
 
       <Divider sx={{ my: '4px !important', p: 0 }} />