+++ packages/ai-runtime/src/types/runtime/schema.ts
@@ -182,6 +182,10 @@ export function outputVariablesToJoiSchema(
       return Joi.any();
     }
 
+    if (variable.from?.type === 'variable') {
+      return Joi.any();
+    }
+
     if (variable.name && isRuntimeOutputVariable(variable.name)) {
       if (variable.name === RuntimeOutputVariable.llmResponseStream) {
         schema = Joi.any();